requirements.txt
numpy==1.15.2
pandas==0.24.2
opencv-python==4.1.1.26
pytesseract==0.3.0
Werkzeug==0.12.2
SQLAlchemy==1.1.13
six==1.10.0
python-editor==1.0.3
python-dateutil==2.6.1
PyJWT==1.5.2
pycparser==2.18
psycopg2==2.7.3
nose==1.3.7
MarkupSafe==1.0
Mako==1.0.7
Jinja2==2.9.6
itsdangerous==1.1.0
ForgeryPy==0.1
Flask==0.12.2
Flask-Testing==0.6.2
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
Flask-Migrate==2.1.0
Flask-Bcrypt==0.7.1
coverage==4.4.1
click==6.7
cffi==1.10.0
certifi==2018.8.24
bcrypt==3.1.3
alembic==0.9.5
urllib3==1.23
idna==2.7
chardet==3.0.4
python-firebase==1.2
pyTelegramBotAPI==3.6.6
python-telegram-bot==11.1.0
requests==2.19.1